requirements.txt
python-dotenv~=1.1.0
huggingface-hub~=0.34.3
streamlit~=1.47.0
langchain~=0.3.25
langchain-community~=0.3.27
langchain-huggingface~=0.3.1
sentence-transformers~=4.1.0
faiss-cpu~=1.11.0post1